+++ src/pages/malynar/register.tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
-import {PageLayout} from '../../src/components/PageLayout/PageLayout'
-import {RegisterForm} from '../../src/components/RegisterForm/RegisterForm'
+import {PageLayout} from '../../components/PageLayout/PageLayout'
+import {RegisterForm} from '../../components/RegisterForm/RegisterForm'
 
 const Register: NextPage = () => (
   <PageLayout seminarId={3}>
