+++ src/pages/malynar/zadania/[series].tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
-import {PageLayout} from '../../../src/components/PageLayout/PageLayout'
-import {Problems} from '../../../src/components/Problems/Problems'
+import {PageLayout} from '../../../components/PageLayout/PageLayout'
+import {Problems} from '../../../components/Problems/Problems'
 
 const ProblemsPage: NextPage = () => (
   <PageLayout seminarId={3}>
