+++ src/pages/strom/gdpr.tsx
