+++ src/pages/strom/register.tsx
