+++ src/pages/strom/vysledky.tsx
